\name{alabama}
\alias{alabama-package}
\docType{package}
\title{Constrained Nonlinear Optimization
}
\description{
Augmented Lagrangian and Adaptive Barrier Minimization Algorithm for optimizing smooth nonlinear objective functions with constraints. Linear or nonlinear equality and inequality constraints are allowed.
}
\details{
\tabular{ll}{
Package: \tab alabama\cr
Type: \tab Package\cr
Version: \tab 2023.1.0\cr
Date: \tab 2023-08-21\cr
License: \tab GPL-2 or greater\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ravi Varadhan (with contributions from Gabor Grothendieck)

Ravi Varadhan <ravi.varadhan@jhu.edu>
}
\keyword{optimize}
\seealso{
\code{constrOptim}, \code{spg}
}
