\name{alabama}
\docType{package}
\title{Constrained nonlinear optimization
}
\description{
Augmented Lagrangian Adaptive Barrier Minimization Algorithm for optimizing smooth nonlinear objective functions with constraints. Linear or nonlinear equality and inequality constraints are allowed.
}
\details{
\tabular{ll}{
Package: \tab alabama\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-24\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ravi Varadhan (with contributions from Gabor Grothendieck)

Ravi Varadhan <rvaradhan@jhmi.edu>
URL:http://www.jhsph.edu/agingandhealth/People/Faculty_personal_pages/Varadhan.html
}
\keyword{optimize}
\seealso{
constrOptim, spg
}
